Moved Client functions in extra file.
Moved functions of the Client object in an extra file to make the wrapper code more readable.

Also implemented the usage of the python library python-box (reference here: https://pypi.org/project/python-box/). Briefly, it is a library to make dicts easier accessible. I used it to make keys in the JSON (returned from LernSax) accessible as attributes.